requirements.txt
customtkinter
SpeechRecognition
pyaudio
python-dotenv
google-generativeai